Rename GuardrailsManager.tsx to GuardrailsApplicator.tsx.
