+++ vscode/webviews/components/GuardrailsApplicator.tsx
